From: "Chris Francisco" <chris@math.okstate.edu> Date: March 7, 2008 9:30:11 AM EST To: "Michael Stillman" <mike@math.cornell.edu> Subject: Re: Macaulay2 march conference
Hi, Mike,

I'm attaching a revised version of LexIdeals, labeled version 1.1. It
has mostly relatively minor changes, but the big change is that I
included a new feature for lexIdeal along the lines of what Eisenbud
wanted. Instead of doing lexIdeal(ring, Hilbert function) to get a lex
ideal with a given Hilbert function, one can now also do lexIdeal of
an ideal, obtaining the lex ideal with the same Hilbert function as
the given ideal. This works for non-Artinian ideals as well as
Artinian. lexIdeal still doesn't support rings that are quotients by
arbitrary homogeneous ideals (i.e., not monomial ideals) because there
are a few extra issues there, and I want to talk to some people at the
conference about it.

I cleaned up the documentation for a few things and added
documentation for the new feature of lexIdeal.


git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@6715 a8012f80-5f13-0410-8dc0-eb77bd940c0d