Update an expected test output file in the complete intersections chapter of the book.
The engine code for gb over fields has been repaired since 1.3.1, so that gb's
(and thus syzygy matrices) are now autoreduced.  This causes the answer
returned by minimalPresentation over polynomial rings to change, because the
algorithm involves reducing matrices modulo all the variables, and using
"complement" over the coefficient field.

Careful examination of intermediate computations in the code Ext(Module,Module)
shows the change in gb to be responsible for the change in the form of the
matrices computed in this test.  The degrees of the generators do not change.

Here's an example:

    Macaulay2, version 1.3.1.1

    i1 : gens gb matrix {{1,1/2},{0,1}}

    o1 = | 1 0 |
	 | 0 1 |

		  2        2
    o1 : Matrix QQ  <--- QQ



    Macaulay2, version 1.3.1

    i1 : gens gb matrix {{1,1/2},{0,1}}

    o1 = | 1 1/2 |
	 | 0 1   |

		  2        2
    o1 : Matrix QQ  <--- QQ


git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@11891 a8012f80-5f13-0410-8dc0-eb77bd940c0d