Bill commented out some needed entries in the "version" hash table, causing the initialization routine in startup.m2.in to fail.
I've put a few of them back in, but with "unknown values".
For example, we need an identifier for the "machine", so that machine
dependent files can be stored in a directory with that name.
And we need to know what version of Macaulay2 this is, so we can
print it out when the program starts up.  And so on.

I'll put the version number in a separate file at top level.  Currently
it's sitting in configure.ac, but if it's in a separate file, configure.ac
can be made to get it from there.  Then Bill can propagate it as a #define
called "MACHINE" for the benefit of the C code.


git-svn-id: svn://svn.macaulay2.com/Macaulay2/branches/gfurnish/scons@14369 a8012f80-5f13-0410-8dc0-eb77bd940c0d