Added infrastructure so that e/unit-tests can test low level code in e dir.
This is not yet complete.  In order to not need to link in everything from d
and libfac, etc, some replacement code has been written (or will be, in the
next commit(s)).  The most important is that global 'new' and 'delete'
replacements from newdelete.cpp are NOT linked in.