Fixed a bug in Hom(--,N) which caused examples to fail.
The code for dual of a matrix uses this code, and it was returning a
matrix which was not evidently a map of free modules, which caused
other tests to fail.  The test 'Macaulay2Doc/test/dual.m2' tests the
behavior of this function, including checking that functoriality
and use of 'homomorphism' still work.