Fixed bug of fmoore (bugs/mike/0-fmoore-pfaffians-res-crash.m2) (part 1)
The bug is one in creating the 'comb' tables, as the memory alloc
in array_class is flawed.  That whole class is flawed, so I rewrote
the comb class, calling it 'Subsets'.  Once 'comb' is eradicated, then
I can delete the array_class class too.

BUT: a bug surfaced that has apparently been a problem for awhile:
exterior power of a Schreyer free module crashes.  Not fixed yet.