This patch disables some non-working examples in the NumericalAlgebraicGeometry package.
When Macaulay2 is linked against newer versions of the mpfr library
(3.1 and higher), it will crash on those examples.  The reason is that
these examples are non-reduced and the algorithms in the NAG package
are not supposed to work on them.

On an up to date Gentoo Linux this makes M2 compile and test fine with
mpfr-3.1.  It seems to me that this was the issue which lead to the
mpfr revert in 32b5fb0ce5 and this revert is now obsolete.