Changed interface of dmat and smat matrix routines.
The interface used to be as a set of template classes, with some
classes specialized.  Now, each function can be specialized by itself
(e.g. determinant, rank).  One has to worry about partial template
specialization of functions.  See
http://www.gotw.ca/publications/mill17.htm
For a good explanation.  Here, we overload, and do not do
partial instantiation of functions.

Only the routines in MatLinAlg have been modified so far, and are now in
mat-linalg.hpp.