fix a bug involving mpz_get_sx
The mpir library has a strange problem that prevents the following short
C++ program from compiling successfully:

    #include <mpir.h>
    #include <stdint.h>
    #include <mpirxx.h>

To get around that, we have been accustomed to patching it, but that
doesn't work when an unpatched mpir library is found on the system and
used.  So here we are careful, instead, to load stdint.h first.