error message was not being handled in GB code
When Hilbert driven GB is being done (or free resolution too),
it can happen that the user gives an incorrect hilbert function.
This is manifested when the expected number of GB elements in a degree
comes out negative (currently, the function hilb::coeff_of takes
caste of this).  In this case, an error is set.  Previously, of the
several calls to this function, most were ignoring this possibility.
Now they all consider it.

Better would be if an error is thrown, and the routines in question were
prepared to handle that exception.  This would be a more serious change.