changed from_int, set_from_int to from_long, set_from_long.
Tests seem to work (except: e/unittests has a problem in RRR.multDivide,
and test/content.m2 isn't fixed yet.
This affected very many files!  A few casts were changed as well.  The
reason I had to change the name: if we have defined from_int(long), and
from_int(mpz_t), then from_int(0) is ambiguous!  0 is an int, and can be
promoted to a long, or a pointer, and it can't decide.