hash codes of symbols
Currently symbols are assigned hash codes sequentially, as part the same
sequence used for mutable objects. But that can make debugging hard, as
the typical way to figure out what a Macaulay2 program is doing is by
editing it to add printing routines, and that often involves creating
new symbols, thus changing the hash codes. So it would be better to
assign symbols a hashcode that depends just on the name of the symbol,
so it wouldn't change so readily.

That's what this commit accomplishes.  It involves adding a new
function, 'serialNumber', to the Core, and some rewriting of the
Serialization package.