Changed behavior of GF not-well-defined ring maps, and GF (Ring).
GF Ring: if the first var is a primitive element (generator of the
multiplicative group, and the size isn't too big (given by SizeLimit),
then an encoding is used.  However, if it is not a primitive element,
then a representation using polynomials is used, and that does not need a
primitive element.

This affects the behavior of ring maps from Galois Fields, that are
not well defined.  A different "lifting" may occur now.  We might
remove the ability to make these not well defined ring maps, but
for now that is still allowed.