add command line flag --no-randomize
and use it for tests and examples (which might change the random number
seed for them).

This is better for debugging in gdb, because merely calling
setRandomSeed based on the time and pid might change the program's
behavior in bad cases.

It would be a good idea to go through all examples, removing random
number seed setting, if it was done just to get the code working.