use libgmp if present
Formerly we insisted on finding libmpir, which can be used to replace
libgmp.  Now we check for libgmp, and if we don't find it, we build
libmpir as its replacement.

I hope this will help get around the problem in issue #155 with
a bug in "ld" under cygwin that has something to do with the
file libmpir.a.