changes needed to compile "statically" under linux 32
Some libraries still need to be dynamic, though:

	ubuntu32$ ldd "/home/dan/src/M2/M2/BUILD/dan/builds.tmp/linux32.generic.production/usr-dist/i686-Linux-Generic/bin/M2-binary"
	linux-gate.so.1 =>  (0xb77d1000)
	libpthread.so.0 => /lib/i386-linux-gnu/libpthread.so.0 (0xb77a8000)
	libdl.so.2 => /lib/i386-linux-gnu/libdl.so.2 (0xb77a3000)
	libgfortran.so.3 => /usr/lib/i386-linux-gnu/libgfortran.so.3 (0xb76a5000)
	libc.so.6 => /lib/i386-linux-gnu/libc.so.6 (0xb74f7000)
	/lib/ld-linux.so.2 (0xb77d2000)
	libgcc_s.so.1 => /lib/i386-linux-gnu/libgcc_s.so.1 (0xb74da000)
	libquadmath.so.0 => /usr/lib/i386-linux-gnu/libquadmath.so.0 (0xb745f000)
	libm.so.6 => /lib/i386-linux-gnu/libm.so.6 (0xb7419000)

It's odd that liblzma is now explicitly required.  I couldn't find a way for the configure script
to determine whether it's required, so for now it's always required.  An improvement to the
configure script would involve trying to link with various libraries statically when searching for
them.