Fixed github issue #203.
The problem was caused by computing a trimmed set of generators of an ideal.  This is done by the GB algorithm using e.g.
  gb (m, StopWithMinimalGenerators=>true);
The GB is designed to stop in this case when a variable n_gens_left is zero.  The logic handling that was incorrect in some cases.  The code was changed to be more robust, and more obviously correct when dealing with this variable.