restore good caching behavior for GBs
Mike noticed that my recent changes to "remainder", part of making Hom
and related functions more general, caused it to stop caching the gb of
the denominator, slowing down example code that calls it repeatedly with
the same denominator.  Here we fix that.

The choice of keys for caching matrices derived from modules is
arbitrary, so feel free to review and change later.