Fix makefiles, now that memtailor uses libtool again (v2)
This commit was previously applied, but then reverted because the
corresponding commit in the memtailor submodule was not updated
(see [1]).

Note that we have updated M2/submodules/memtailor to point to [2], one
commit earlier than the current HEAD of the master branch.  It may be
preferable to merge [3] first and use that commit instead.

[1] https://github.com/Macaulay2/M2/pull/343
[2] https://github.com/Macaulay2/memtailor/commit/a8212c6
[3] https://github.com/Macaulay2/memtailor/pull/3