Added Pullback.m2  This has a function for computing a pullback in the category of rings.  This is very useful for constructing non-normal rings and in the future might also be used to help write seminormalization and weak normalization algorithms.
We also include an updated PushForward.m2  Claudiu Raicu asked me to update this here too.  This has a change where the target of the third output from pushing forward a module is the actual desired module and not just a free module.
It also includes an option for turning off prune calls in the pushFwd function.