Major Revision to all methods
Added the capability to work with subschemes of any product of projective spaces as well as subschemes of some toric varieites as described in http://arxiv.org/abs/1508.03785. Also added new options for increased user control over the internal computations and increased options for what computed values will be retained.