Fix a miscalculation of the expected string length
In a debug build, concatenate("ABC",-1) will crash M2 because it miscalculates
the expected length of the final string (concatenate("ABC",2) gives "ABC  ").
Negative numbers like this arise by doing something like pad("ABC",2);
though this change makes it just return the whole string, reasonable behavior could also
be to truncate the string.

This was causing a crash in the "minprimes" example inside the MinimalPrimes package.