Check denom!=0 when fraction is converted to ZZ/p
The code "substitute(1/101,ZZ/101)" currently returns 0 on a
non-debug build, and terminates in ARingZZpFlint::divide with
"Assertion `b != 0' failed" in a debug build.  Compare that to
"promote(1/101,ZZ/101)" which gives "error: promotion of this
rational number to the ring R not possible" on both types of
builds.

Change the behavior of ARingZZpFlint::set_from_mpq to error
if the denominator was zero, before it tries to do division in
ZZ/p.  I'm not positive I'm handling the error correctly.

This was noticed because packages/Macaulay2Doc/test/subst5.m2
was failing on debug builds.