Use gftables from factory if factory is not built.
The configure script checks in the expected location for the gftables
directory or it may be specified using GFTABLESDIR.  Note that we must
expand the ${datadir} variable recursively in order for AC_CHECK_FILE
to work.  For this, we use the adl_RECURSIVE_EVAL macro from the Autoconf
Archive [1].

This fixes #375.

[1] http://ac-archive.sourceforge.net/adl/relpaths.html