bWriteInputFile
Renamed makeB’InputFile to bWriteInputFile.
The main goal is to eliminate the use of apostrophe’s in the name of
files because they can be confused with similar punctuation.
Another goal is to rename functions to start with ‘b’ or ‘bertini’ to
be consistent and to change ‘make’ to ‘write’.