bertiniUserHomotopy is a new function
bertiniTrackHomotopy is a user homotopy of type 1.
The functions here do not have to be polynomials. We should account for
this.
bertiniUserHomotopy is a user homotopy of type 2.
The polynomials here can be defined over product spaces (Use
AffVariableGroup and HomVariableGroup). The parameters must be
specified in terms of a path variable.