Addition of two new packages
The package StronglyStableIdeals has bee submitted to JSAG
in 2014, revised in 2015 according to referee's comments,
but no more volumes of JSAG have been published since then.
The main feature is a method to compute the complete list of
saturated strongly stable ideals given the number of
variables and the Hilbert polynomial.
The brief article prepared for JSAG and explaining the
contents of the package is available at
https://arxiv.org/abs/1406.6924.

The package SLnEquivariantMatrices is an ancillary file to
the paper "A construction of equivariant bundles on the space
of symmetric forms" recently appeared on arXiv
(see https://arxiv.org/abs/1804.06211).
It allows to compute spaces of matrices of constant rank
that are SL(n)-equivariant.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch release-1.11
# Your branch is up to date with 'origin/release-1.11'.
#
# Changes to be committed:
#	new file:   M2/Macaulay2/packages/SLnEquivariantMatrices.m2
#	new file:   M2/Macaulay2/packages/StronglyStableIdeals.m2
#
# Changes not staged for commit:
#	modified:   M2/INSTALL
#	modified:   M2/Macaulay2/packages/=distributed-packages
#	modified:   M2/submodules/fflas_ffpack (untracked content)
#
# Untracked files:
#	M2/Macaulay2/m2/.gdbinit.paolo
#	M2/libraries/cohomcalg/Makefile
#	M2/libraries/topcom/Makefile
#