change configure script so libgc is always rebuilt
We always build gc now, because the homebrew bdw-gc package (at least) builds
the library in such a way that running the tests in the package
"CoincidentRootLoci" results in several crashes with the message "Too many
heap sections: Increase MAXHINCR or MAX_HEAP_SECTS".  Perhaps our use of the
configure option "--enable-large-config" is what makes the difference.