Find system gftables if they exist.
d5bf979 introduced a regression in the feature added in 68a00e7.  In
particular, if we don't build factory, then we should use the gftables
already present on the user's system.

So we looked in two paths -- one prepended with `prefixDirectory`, for
M2-built factory, and one not prepended with `prefixDirectory`, for
system factory.

The second option was removed, an so we looked for gftables only in the
build directory.  But this would fail if we were using system factory.

This patch restores both options.