various improvements
We stop redirecting operator new to libgc in e/newdelete.cpp, as well is
configure.ac, where we recently added -lgccpp.

It was crashing because we no longer redirect operator new to libgc, and Bill's
code in system/ was using operator new to store pointers to gc-allocated
memory.  We added explicit libgc allocation there.

We emptied the contents of scc_core_prepare(), since it seems not to be useful.
It's also no longer a constructor.

We ensured that
    #define GC_FREE_SPACE_DIVISOR 12
    #define GC_INITIAL_HEAP_SIZE 70000000
are obeyed by GC_INIT, which is now run in main().

We eliminated some warnings from gcc about M2lib.c related to the use of "const".

We told pari *not* to initialize the gmp memory functions.