avoid calling GC_throw_bad_alloc()
GC_throw_bad_alloc() is in libgccpp, which we do not want
to link with, as doing this places `::new` under GC's control.
That is, instances allocated by `::new` become traceable by the
GC (although they are only collected if the corresponding class
derives from the gc class).

See https://github.com/antonleykin/M2/pull/5 for a discussion.