fix interrupt handling
The command line option "--int" is supposed to arrange for interrupts not to be
handled, but it wasn't working.  This is important when running examples and
tests, because if we interrupt one, we want to know about it.

I don't know when the code was broken.

Also fix some warning messages about assigning to "const" variables.