fix interrupt handling, part 3
One further problem: on some systems the "system" function uses the shell
/bin/sh to run the command, if it detects an interrupt occurred in the command,
it adds 128 to the interrupt number and returns that as a status code, so the
number returned by "run" may be wrong.

We also institute a uniform treatment of this issue for the Macaulay2 functions
"run" and "wait".