Don't bother calling cmake when installing googletest.
We don't need to actually build it, as it's already built by memtailor,
mathic, and mathicgb during their builds.  They just need to know where
the source is.  And by only running cmake, we weren't even building it
anyway -- we were just preparing to build it.

Since we don't need cmake anymore, we remove the information about
installing it in various Linux distros from INSTALL.