Print useful message and raise an error when call to gfan fails
The printing of a message existed in the code but was commented out.
In addition, we raise an error.  If our call to gfan doesn't work,
then the user should be informed right away rather than encountering
a less useful error a bit later when Macaulay2 isn't able to deal
with the output that doesn't exist.

Closes: #1066