Properly restore missing diagrams from MonomialAlgebras package.
In the previous attempt (#1131), an absolute path using prefixDirectory
was used for the src attribute of each img tag.  This was hardcoded into
the documentation during build, and so the image links were still broken.

We now skip using prefixDirectory and let html.m2 compute the proper
relative path to the images.