Get the source code for functions defined in startup.m2.in.
This file only exists in the Macaulay2 source code and is not installed.
However, its contents are available as a string, and we can still
access it.  The code to do this exists in code.m2, but was still
set up for when these functions were defined in layout.m2, which hasn't
existed since fb2a04e.

We also restore the SoureCode option to applicationDirectory's
documentation.  It was removed in 728c632 because of this bug.