Fix bug in installPackage when MakeDocumentation => false
Currently, we get the following:

```
i1 : installPackage("FirstPackage", MakeDocumentation => false)
--installing package FirstPackage in /home/profzoom/.Macaulay2/local/ with layout 1
--using package sources found in /home/profzoom/tmp/
stdio:1:1:(3): error: no method for binary operator | applied to objects:
--            null (of class Nothing)
--      |     ".installed" (of class String)

```

The `null` is coming from the `libDir` variable, which is only defined
when `MakeDocumentation` is true.  We move the assignment of this variable
(which is only used once) to immediately before it is used.