Revert commits that changed behavior to display relative paths
The goal of these commits was to prevent the build directory from appearing
in the documentation for reproducible builds.  However, the commit which
made this work (building the examples in a subdirectory of the build
directory instead of /tmp) was causing build failures and has been
reverted.  Without that commit, the build directory still appears, but as
../../../build/path instead of /build/path.