Use generic "/path/to/source/" path for cached files
These files are not installed with Macaulay2, and instead their contents
are cached as strings inside the Macaulay2 binary.  Currently, the
original source directory is displayed when running the basic tests
or locating the code of a function defined in startup.m2.in, which results
in non-reproducible builds and also may be confusing to users.

We instead replace the beginning of these paths with "/path/to/source/"
to indicate that these files are only present in the source code.