undo a change designed to make builds reproducible
... because, by setting the environment variable HOME to a fictional value,
it prevents shell scripts that need to find the user's files using the HOME
directory from working.

In this case, the shell script was the one I installed to execute the external
program "bergman".  Here it is:

    #! /bin/bash
    export BERGMANPATH=$HOME/src/bergman1.001
    export bmload=$BERGMANPATH
    export bmroot=$BERGMANPATH
    exec $BERGMANPATH/bin/clisp/unix/bergman "$@"

How could the shell script find the program without a true value for HOME?