Turn on factory's extended Zassenhaus GCD switch for finite fields
Without it, factory 4.1.1 hangs when factoring large polynomials (e.g., the
one at the end of the `timing-quotient.m2` test.  It is already switched
on by default in factory >= 4.1.2.

This was originally fixed by patching factory itself (see #783), but by
turning on the switch in Macaulay2, we can use factory provided by the
system.  Since it is redundant now, we also remove the corresponding line
from the factory patch.

Closes: #1316