Fix check-configure so that it works whether we build factory or not
A recent change "fixed" this check so that it worked if we were using
the gftables from a pre-existing installation of factory.  However, in
the process it broke the check if we build factory.  In particular, it
was looking for them in the final installation location and not in the
staging area, where they are locating during this check.

To fix this, we add a new variable, PRE_GFTABLESDIR, which stores the
location of the gftables during the check.