modify the documentation of PHCpack
so the example output does not vary from run to run.  This makes things simpler
for those building Macaulay2 on a machine where phc is installed, for then the
example output files do not get modified.  The main difference is that the use of
Verbose=>true is avoided, for that displays temporary filenames, which change
every time.