Use TO instead of TO2 when listing documentation templates
This improves the experience for users viewing the "writing
documentation" node.  Previously, the names of these templates as
given in the list were slightly different than the corresponding
documentation keys.  This was not a problem for users viewing the html
or info documentations thanks to hyperlinks, but it was potentially
confusing for users using "help".

Now when using "help", the names of the templates will be enclosed in
quotes, providing a visual cue that these are documentation nodes.
Users may also take advantage of the "* String" method to easily pull
up the documentation.