Add support for program prefixes in loadProgram
Frequently, distributions add prefixes to the many binaries which come
with a particular program, e.g., 4ti2-, nauty-, and topcom- in Debian.
We add a Prefix option to loadProgram which takes a list of triples:
(name of distribution, regex to match binaries with the prefix, prefix).
Each one is tried (in addition to the default with no prefix)
and the triple that works is stored in the Program object for later use.