Use multiple assignment for an example in PushForward
The pushFwd method on RingMap objects returns a sequence containing
three elements.  We rewrite the example to look at each of these
elements individually.

In particular, the final element is a function, and previously only
the location of the function was displayed in the example.  This
led to reproducibility issues as the path might be split by wrapping,
making it difficult for the reproduciblePaths function to replace
the build path.  We now demonstrate the use of the function instead,
so the path is no longer even displayed.