Link against Boost.Stacktrace in autotools build
As discussed in #1429, this provides much more informative backtraces and
is already being done in the CMake build.  We first try linking against
libboost_stacktrace_backtrace, which is available in Ubuntu but not OS X,
and then libboost_stacktrace_addr2line, which is available in OS X. If
neither are found, then we go back to the original behavior of using
the header-only library.