fix a bug in chain complexes
The problem was that the method for Hom(ChainComplex,Module) would scan only
the differentials of the chain complex and not the modules.  If a module had
nothing adjacent to it, then it wouldn't be accounted for among the
differentials, and so the dual module wouldn't get inserted, resulting in a
chain complex with missing modules.