fix printing in threads, part 1
some assertion statements have been put into stdio.d

Currently working on getting this code to work:

   n = 0 ; schedule( () -> while true do (n=n+1; << "hi there hi there hi there hi there hi there hi there  " << n << endl << flush ));

The problem seems to be that the FOSS obtained by the master thread and the scheduled thread is the same, so somehow
the "outindex" field in it is altered by the master thread when it prints the next prompt.  That happens after the scheduled
thread has printed its message about 500 times, and thread context switching most often occurs while the scheduled thread
is in the "write" system call.

The next step would be to check whether the two FOSS structures really are the same, perhaps by writing a little code so the
address can be printed from top level.