Modify readNmzData so that it calls readMultipleNmzData
The former is just a special case of the latter, so now we only need to
implement the matrix-parsing algorithm once.

With the changes to readMultipleNmzData in the previous commit, this
package now supports any halfway recent version of normaliz regardless
of whether its output includes the numbers of rows and columns in a matrix
or not.  Closes: #1439