Implement --check 2 and --check 3 command line options.
According to `M2 --help`, these options are supposed to:

```
    --check n          run tests to level n
                           n=1: basic tests
                           n=2: test Core
                           n=3: test all packages
```

However, only `--check 1` is currently implemented.  We implement the
others by running "check" for "Macaulay2Doc" only for `--check 2` and
for *all* packages for `--check 3`.  Afterwards, Macaulay2 exits, and
the exit code is equal to the number of failing packages.

This will make running continous integration tests on installed copies of
Macaulay2 simpler.