Refactor check methods
The code for `check Package` and `check(ZZ, Package)` were essentially
the same.  We combine them by making it so that if the latter is
called with an argument of -1, then all of the tests are run, and so
the former just calls this.

We also keep track of the specific indices of tests with errors for an
improved error message.