Only require one variable for verbosity in autotools build
Previously, the autotools build had two variables controlling the
verbosity of the build:

* `VERBOSE`: See `M2/include/config.Makefile.in`.  Controls the output during
  compilation, e.g., `CXX foo.cc` v.
  `g++ (lots of build flags) foo.cc -o foo.o`.

* `Verbose`: See `M2/Macaulay2/packages/Makefile.in`.  Controls the
  output when building examples and testing packages.  In particular,
  whether to display what happened when these fail.

We make it so that only the former is necessary.  This brings the
autotools build in line with the behavior of the CMake build, which
uses `$<IF:$<BOOL:${VERBOSE}>,true,false>` to determine the value of
`Verbose`.