Find location of failing test output correctly when running a single test
`temporaryFilenameCounter` is incremented twice every time we run a
test, so we use this fact to compute the location of the temporary
file containing the output of failing tests when the `Verbose` option
to `check` is `true`.

However, this was done incorrectly when a single test was checked.  In
particular, the number of the test was being used to compute the
location of the temporary file, but this number is meaningless in this
case.  Instead, we always want to subtract 2 from the current value of
`temporaryFilenameCounter`.

Closes: #1686